requirements.txt
pillow==10.0.1
pandas==2.0.1